Fix annotations in year 2038
Fixes #3050

+++ src/Task.cpp
@@ -1189,7 +1189,7 @@ void Task::addAnnotation (const std::string& description)
 
   do
   {
-    key = "annotation_" + format ((int) now);
+    key = "annotation_" + format ((long long int) now);
     ++now;
   }
   while (has (key));
